✨ Expand / correct attachmentVersion value type
The value field for attachmentVersion incorrectly required the value to
be a cost value type. This commit expands the boilerplate to allow for
any attachment type, and use t.unknown for types that are not yet
known about.

This does not ensure that the value type matches what should be expected
given the attachment type (given that this data is contained within
another table), bit it at least ensures that the value type is ONE of
the expected types.