Define bigint columns as a union of string and number
Idea to use BigInt type for columns defined as int8 (bigint)
was a good one, but requires pg serialization changes, which
affect old sequelize models, not only new knex models.

This switch would not be a walk in the park, especially since
there are lots of legacy code working with financial data,
which would suddenly need to be differently typed.

This change would require much more careful planning and
testing than initially thought, so revert it is.

Union of string and number is picked because int8 and numeric
DB types are read as strings, but when inserting rows, we don't
want library clients to provide numbers as strings.