✨ Introduce new custom where condition construction
Given a knex query builder, and one of our own custom conditions,
return the prepared condition to be used directly by knex.
There are a few reasons we do this rather than simply using knex query
builder directly:

* the query builder does not offer good type-safety (for example whereIn
  does not require the type of the values to match the type of the column)
* we find ourselves needing to use query builders often for very simple
  conditions that could be abstracted + simplified

However, we retain the ability to use a query builder when needed.