✨ HPC-8267: Ignore inconsistent project budgets
It is quite common (and expected) to encounter inconsistent project
budget data when looking at incomplete / draft / un-submitted projects,
for example in the gap-analysis endpoint.

To allow for these situations, we have introduced a flag that lets
these inconsistencies be ignored. Projects that have inconsistencies
will not have associated data returned as part of the result when this
flag is true.