Search participants by `hidSub` directly from DB
We currenly search the `participant` table by `hidSub`
column, expecting 0 or many results. Then, `Array.prototype.filter()`
is applied on the results, effectively applying
the same condition by which we queried the DB,
so we get the first participant with a given `hidSub` back.

This overhead is not necessary, since `hidSub` is an
unique column (has `UNIQUE INDEX`). This code is a
leftover from days when `hidId` and `hidSub` coexisted
in `participant` table and were both used for querying,
where one got the precedence over the other -
https://github.com/UN-OCHA/hpc_service/commit/da4256a340847495237e280cf2f5a72bba3aaca7

We switched to querying `hidSub` exclusively in
https://github.com/UN-OCHA/hpc_service/commit/a8e027f8e6b01e1af7b71befb4a84f5b4be40b61, but this
pattern of using `Array.prototype.filter()` persisted.