HPC-8748: Introduce project deleting util method
When a project is being deleted, a bunch of
pre-requisites need to be fulfilled. Then,
in case that deletion can successfully go
through, we also need to clean up any references
to the project, especially those which are
not cleaned up automatically by DB's
`ON DELETE CASCADE` foreign key constraints.