HPC-9460: Allow for read-only replica DB connection
Since Knex doesn't support read-only replicas,
we need to allow for a second connection to
the replica server. This assumes master and
replica DB are on separate servers, but if
connection to replica isn't passed, the models
behave just like before.

The goal is to use master DB for write operations
and for transactions and replica DB should be
only used for read-only operations. There are
lots of changes, but it's mostly passing around
of connections, just so that newly introduced
replica connection can be used in `find()`
method of raw model.