Cache conversion to camel case
The method `toCamelCase` proved to be a bottleneck
because it is used to convert strings in disaggregated
data and some plans have enormous amounts of data,
thus, in case of caseload dissagregation of plan
with ID 1185, this caching reduced number of calls
from 16,850,784 to just 4,584! Effectively, this
reduces the time from ~25 seconds to less then 5,
in tests on a local environment.