Return array instead of `Set` from `groupObjectsByProperty()`
This is done because `Set` will treat objects the same only if
> both are the same object (meaning both values reference the same object in memory)

This can be illustrated with an example:
```
new Set([{ duplicate: true }, { duplicate: true }]);
// Gives Set(2) {{…}, {…}}
```

The two objects are considered as different, even
though they are same by value. As said, uniqueness
of objects is only if they use same reference.
```
const obj = { duplicate: true };
new Set([obj, obj]);
// Gives Set(1) {{…}}
```

Therefore, to break the illusion of uniqueness that
is logically associated with Sets, the grouping is
changed to arrays. This enables using `Map.groupBy()`