Use `Iterator.prototype.map()` directly
Since Node.js 22, we have iterator helper methods
like `map()` which we can use to iterate `Map`, `Set`
and other `Iterator` instances, without the need
to convert to array first.

But, a big caveat is that these iterators are lazy
and can be consumed only once, so we need to beware!